# fe-pickup/openapi.yaml
openapi: 3.1.0
info:
  title: PickUp API
  description: PickUp API 문서
  version: v1
servers:
- url: /
...
paths:
  /members:
    post:
      tags:
      - Member
      summary: 회원가입
      description: "로그인 ID, 닉네임, 비밀번호로 회원을 생성합니다."
      operationId: createMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRequest"
        required: true
      responses:
        "201":
          description: 회원가입 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "409":
          description: 로그인 ID 또는 닉네임 중복
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
...
  /members/me/watches:
    get:
      tags:
      - Member
      operationId: getMyWatches
      summary: 관심 목록 조회
      description: |
        관심 등록한 예정(SCHEDULED) 경매만 최신순으로 조회합니다. 낙찰되면
        관심 목록에서 노출되지 않습니다. 다음 페이지 조회 시 이전 응답의
        cursor를 전달합니다.
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: 관심 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyWatchesPageResponse"
        "400":
          description: 유효하지 않은 커서 값
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
  /members/me:
    get:
      tags:
      - Member
      summary: 내 정보 조회
      operationId: getMyProfile
      responses:
        "200":
          description: 내 정보 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MyProfileResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 회원 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
    patch:
      tags:
      - Member
      summary: 내 정보 수정
      description: "전달된 닉네임, 비밀번호, 프로필 이미지만 수정합니다. 프로필 이미지는 SET 또는 REMOVE로 변경하며 비밀\
        번호 변경 시 현재 비밀번호가 필요합니다."
      operationId: updateMyProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMyProfileRequest"
        required: true
      responses:
        "200":
          description: 내 정보 수정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MyProfileResponse"
        "400":
          description: 요청 값 검증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 회원 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "409":
          description: 닉네임 중복
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
...
  /members/me/wins:
    get:
      tags:
      - Member
      summary: 내 낙찰 내역 조회
      description: 회원이 낙찰받은 경매 목록을 최근 입찰 순으로 조회합니다. 경매당 마지막(최신) 입찰이 낙찰(WON)인 것만 반환합니다.
      operationId: getMyWins
      parameters:
      - name: getMyWinsRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/GetMyWinsRequest"
      responses:
        "200":
          description: 내 낙찰 내역 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MyBidsPageResponse"
        "400":
          description: 유효하지 않은 커서 값
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
  /members/me/points:
    get:
      tags:
      - Member
      summary: 내 포인트 잔액 조회
      operationId: getMyPointBalance
      responses:
        "200":
          description: 포인트 잔액 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PointBalanceResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 포인트 정보 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
  /members/me/point-transactions:
    get:
      tags:
      - Member
      summary: 내 포인트 거래내역 조회
      description: 실제 포인트 증감 내역을 최신순 커서 기반으로 조회합니다.
      operationId: getMyPointTransactions
      parameters:
      - name: getPointTransactionsRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/GetPointTransactionsRequest"
      responses:
        "200":
          description: 포인트 거래내역 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PointTransactionsPageResponse"
        "400":
          description: 유효하지 않은 커서 또는 페이지 크기
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
  /members/me/point-charges:
    post:
      tags:
      - Member
      summary: 포인트 충전
      description: |
        실제 결제 연동 없이 목업으로 포인트를 즉시 적립합니다. 같은 idempotencyKey로 재요청하면
        중복 적립되지 않고 이전 결과를 그대로 반환합니다.
      operationId: chargeMyPoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargePointRequest"
            examples:
              포인트 충전 요청:
                description: 포인트 충전 요청
                value:
                  amount: 300000
                  idempotencyKey: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        required: true
      responses:
        "201":
          description: 포인트 충전 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PointChargeResponse"
        "200":
          description: 이미 처리된 충전 요청(같은 idempotencyKey 재요청)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PointChargeResponse"
        "400":
          description: 유효하지 않은 충전 금액
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "404":
          description: 회원 없음
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
  /members/me/bids:
    get:
      tags:
      - Member
      summary: 내 입찰 내역 조회
      description: 회원이 입찰한 경매 목록을 최근 입찰 순으로 조회합니다. 경매당 마지막(최신) 입찰만 반환합니다.
      operationId: getMyBids
      parameters:
      - name: getMyBidsRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/GetMyBidsRequest"
      responses:
        "200":
          description: 내 입찰 내역 조회 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MyBidsPageResponse"
        "400":
          description: 유효하지 않은 커서 값
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
        "401":
          description: 인증 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionResponse"
      security:
      - access-token-cookie: []
...
components:
  schemas:
    MemberResponse:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        loginId:
          type: string
        nickname:
          type: string
        profileImageUrl:
          type: string
    ExceptionResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
    MemberRequest:
      type: object
      properties:
        loginId:
          type: string
          maxLength: 2147483647
          minLength: 4
        nickname:
          type: string
          maxLength: 2147483647
          minLength: 4
        password:
          type: string
          maxLength: 2147483647
          minLength: 4
      required:
      - loginId
      - nickname
      - password
...
    SearchCardsResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        language:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
...
    MyProfileResponse:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        loginId:
          type: string
        nickname:
          type: string
        profileImageUrl:
          type: string
    ProfileImageUpdateRequest:
      type: object
      properties:
        action:
          type: string
          enum:
          - SET
          - REMOVE
        temporaryObjectKey:
          type: string
      required:
      - action
    UpdateMyProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 2147483647
          minLength: 4
          pattern: .*\S.*
        currentPassword:
          type: string
          maxLength: 2147483647
          minLength: 4
        password:
          type: string
          maxLength: 2147483647
          minLength: 4
          pattern: .*\S.*
        profileImageUpdate:
          $ref: "#/components/schemas/ProfileImageUpdateRequest"
...
    GetCardDetailResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        cardName:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        language:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
...
    MyBidListItemResponse:
      type: object
      properties:
        auctionId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/GetCardDetailResponse"
        grade:
          type: string
        myBidPrice:
          type: integer
          format: int64
        currentPrice:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - HIGHEST
          - OUTBID
          - WON
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
    MyBidsPageResponse:
      type: object
      properties:
        hasNext:
          type: boolean
        cursor:
          type: string
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/MyBidListItemResponse"
    WatchItemResponse:
      type: object
      properties:
        auctionId:
          type: integer
          format: int64
        consignmentId:
          type: integer
          format: int64
        card:
          $ref: "#/components/schemas/SearchCardsResponse"
        grade:
          type: string
          nullable: true
        auctionStatus:
          type: string
          enum:
          - SCHEDULED
          - ONGOING
          - WON
          - PASSED
        startingPrice:
          type: integer
          format: int64
        currentPrice:
          type: integer
          format: int64
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        remainingSeconds:
          type: integer
          format: int64
          nullable: true
        watchCount:
          type: integer
          format: int32
        watched:
          type: boolean
        thumbnailUrl:
          type: string
          nullable: true
    MyWatchesPageResponse:
      type: object
      properties:
        hasNext:
          type: boolean
        cursor:
          type: string
          nullable: true
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/WatchItemResponse"
...
    GetMyWinsRequest:
      type: object
      properties:
        cursor:
          type: string
        size:
          type: integer
          format: int32
    PointBalanceResponse:
      type: object
      required:
      - pointBalance
      - reservedPointBalance
      - availablePointBalance
      properties:
        pointBalance:
          type: integer
          format: int64
        reservedPointBalance:
          type: integer
          format: int64
        availablePointBalance:
          type: integer
          format: int64
    GetPointTransactionsRequest:
      type: object
      properties:
        cursor:
          type: string
        size:
          type: integer
          format: int32
    PointTransactionItemResponse:
      type: object
      required:
      - pointTransactionId
      - transactionType
      - amount
      - balanceAfter
      - createdAt
      properties:
        pointTransactionId:
          type: integer
          format: int64
        transactionType:
          type: string
          enum:
          - OPENING_BALANCE
          - AUCTION_PAYMENT
          - AUCTION_PAYOUT
          - CHARGE
        amount:
          type: integer
          format: int64
        balanceAfter:
          type: integer
          format: int64
        auctionId:
          type: integer
          format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
    PointTransactionsPageResponse:
      type: object
      required:
      - hasNext
      - size
      - items
      properties:
        hasNext:
          type: boolean
        cursor:
          type: string
          nullable: true
        size:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/PointTransactionItemResponse"
    ChargePointRequest:
      type: object
      required:
      - amount
      - idempotencyKey
      properties:
        amount:
          type: integer
          format: int64
        idempotencyKey:
          type: string
    PointChargeResponse:
      type: object
      required:
      - pointTransactionId
      - chargedAmount
      - pointBalance
      - reservedPointBalance
      - availablePointBalance
      - createdAt
      properties:
        pointTransactionId:
          type: integer
          format: int64
        chargedAmount:
          type: integer
          format: int64
        pointBalance:
          type: integer
          format: int64
        reservedPointBalance:
          type: integer
          format: int64
        availablePointBalance:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    GetMyBidsRequest:
      type: object
      properties:
        cursor:
          type: string
        size:
          type: integer
          format: int32
...
  securitySchemes:
    access-token-cookie:
      type: apiKey
      description: 로그인 시 발급되는 access-token 쿠키
      name: access-token
      in: cookie